Remote execution: add support for strong consistency, chunk info logging and other features that rely on HTTP headers (#12745)
#### What this PR does

This PR adds support for strong consistency to the remote execution path
introduced in https://github.com/grafana/mimir/pull/12551 and
https://github.com/grafana/mimir/pull/12302.

I've introduced the concept of `Injector`s, which are responsible for
adding headers / metadata to a request from query-frontends to queriers,
and the concept of `Propagator`s, which are responsible for extracting
information from headers / metadata and adding them to the
`context.Context`.

This model also allows us to support other features in GEM that relied
on propagating information in HTTP headers, so I've reverted #12699.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.