MQE: support delayed name removal with remote execution (#12687)
#### What this PR does

This PR adds support for delayed name removal to remote execution.

The state of delayed name removal is now attached to the query plan,
rather than relying on the configured options in the `Engine` instance.
This ensures consistent behaviour while rolling out delayed name removal
(or disabling it) - otherwise some queriers could behave as if delayed
name removal is enabled, and some could behave as if delayed name
removal is disabled.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12551 and
https://github.com/grafana/mimir/pull/12509

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.