continuous-test: Improve record test logging/metrics, adjust tuneables (#12754)
#### What this PR does

This PR:
- Removes per-record statistics logs (anything per-record can generate
spam) and replace them with metrics where appropriate.
- Tunes tuneables:
- The limit at which to skip a partition on the current run and just
jump to the end. This was simply too small for large cells and we jump
too often.
- Reduces `fetchMaxBytes` by 10x, to force more smaller fetches. Cells
with a high partition count still grab a staggering number of records
per fetch. Even with limits, this doesn't take advantage of parallelism
very well.

#### Which issue(s) this PR fixes or relates to

contrib: https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.