Add Compactor Scheduler Skeleton (#12435)
This PR introduces experimental support for a "compactor scheduler" mode
that allows the compactor to request jobs from an external scheduler
service. This PR makes a few changes to the `MultitenantCompactor` but
filters most of the logic for the new mode through a new concept, a
`CompactionExecutor`.

The `CompactionExecutor` currently has two implementations,
`StandaloneExecutor` and `SchedulerExecutor`. The former is a copy of
the traditional compactor's `running` method, the latter implements the
experimental execution pattern.

To support the new mode, this PR adds the following configuration
options :
* `-compactor.mode` Sets operation mode (one of `"standalone"`,
`"scheduler"`)
  * `-compactor.scheduler-endpoint` - Scheduler service address
* `-compactor.scheduler-update-interval` - How often does the compactor
send an update to extend its lease on a job with the scheduler endpoint.

As of writing there are a few caveats:
- This PR assumes the scheduler is always up - or the penalty for it
going down is trivial. I don't think this is a reasonable long-term
assumption.
- We may consider a job history snapshotting mechanism (see:
`BlockBuilderScheduler, snapOffsets` in `pkg/blockbuilder`) or retries
on specific RPCs (e.g. `SchedulerExecutor.updateJobStatus`)

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Jonathan Halterman <jonathan@grafana.com>