Update mimir-prometheus to 14d2a30d6abe (#12768)
This PR is a fix applied to the automatically generated PR #12764. The
fix decodes histogram buckets when a query contains a subquery since we
need to handle counter resets correctly.

Specifically, it's applying the same treatment to MQE as the change in
commit
https://github.com/prometheus/prometheus/pull/17127/commits/0746f388b047e130e6864d145bb18c47483b3ad8
from https://github.com/prometheus/prometheus/pull/17127