Remove unnecessary `subtreeFolder` (#12785)
#### What this PR does

This PR removes `subtreeFolder`. As far as I can tell, `subtreeFolder`
is not needed:

If all legs of the expression are shardable, then `shardSummer` will
rewrite the expression with magic `__embedded_queries__` selectors, and
if the expression is not shardable, then `shardSummer` will leave the
expression as-is.

`subtreeFolder` then wraps any unshardable expression with a
`__embedded_queries__` selector.

But, if the expression isn't shardable, then the query sharding
middleware ignores the expression produced by `shardSummer` and
`subtreeFolder` and evaluates the original expression as-is in queriers
([source](https://github.com/grafana/mimir/blob/8d60752027b7263667d4492855bf8b698d11699a/pkg/frontend/querymiddleware/querysharding.go#L126)).

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.