inex planning: Add statistics generation for head block (#12752)
## Summary

This PR implements a background statistics service that pre-computes
query planning statistics for head blocks. The idea is to move all
statistics collection and creation out of prometheus and inside mimir.
This is the last part of https://github.com/grafana/mimir/issues/12604.

This includes the changes from
https://github.com/grafana/mimir-prometheus/pull/981


## Necessary changes

* the metrics we register are no longer in the prometheus TSDB; instead
we instantiate a per-tenant `lookupplan.Metrics` instance for every
tenant

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>