Remove `MimirFrontendQueriesStuck` alert (#12810)
#### What this PR does

As of https://github.com/grafana/mimir/pull/12200 (part of
https://github.com/grafana/mimir/issues/11884), the query-scheduler is
required, so the `MimirFrontendQueriesStuck` alert is no longer
relevant.

The `MimirSchedulerQueriesStuck` alert is the equivalent alert for the
query-scheduler and has not been removed.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/issues/11884

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.