Add support to configure max eventual consistency delay on a per-query basis (#12751)
#### What this PR does

In this PR I'm adding the experimental support to configure the max
eventual consistency delay on a per-query basis. The max delay is
currently propagated via a new internal HTTP header named
`X-Read-Consistency-Max-Delay`, but in a follow up PR I will try to
reuse the existing `X-Read-Consistency` header to both specify the level
(as of today) and the max delay (new).

The max delay is intended to be used internally, and in particular by
ruler. When a rule is evaluated with eventual consistency, we may want
to specify a max delay to ensure it doesn't get executed on very stale
data. I initially though to set the delay equal to the configured
`ruler_evaluation_delay_duration`, but in this PR I'm introducing a new
experimental configuration option called
`ruler_evaluation_consistency_max_delay`. The new config option give us
the ability to experiment with the delay, before deciding whether just
setting the delay to `ruler_evaluation_delay_duration` is a good idea or
not.

When the ingester receives a request with `eventual` consistency level
and max delay configured, the ingester checks if the max delay is
honored. If it's not honored, the ingester immediately fail the query. I
opted for this solution, rather than having the ingester waiting until
max delay is honored, because the idea is to set max delay to a fairly
high value (e.g. 1 minute) and so waiting extra time may not give much
benefit, but would slow down other enqueued queries not having max delay
enforcement (the possible slowdown is due to saturating querier workers,
that would get occupied just by waiting for max delay to be enforced).

The max delay is not enforced when the consistency level of a query is
`strong`, because strong consistency already implies an up to date
state.

I've also manually tested this PR in a local dev env, injecting a fake
high latency when consuming from Kafka and looks working as expected.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>