store-gateway: use streaming for skipchunks requests
This changes store-gateways to use the streaming code path for requests
where chunks are not sent. This provides the same results to queriers
since they already know how to parse the streaming response and allows
us to remove the non-streaming code path in the store-gateway.

Notably however, this change does not remove any of the non-streaming
code from queriers. This is required to avoid failing requests during
a rollout. When queriers make requests without chunks to new store-
gateways, the streaming code path will be used and queriers will be
able to parse the result. When requests without chunks are made to
older store-gateways the non-streaming code path will be used which
queriers are still able to parse the result of.

Part of #12673

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>