Avoid continuous active series tracker reloads when cost attribution is in overflow (#12822)
#### What this PR does

This fixes the continuous (every cooldown period) reloads of the active
series trackers when the cost-attribution tracker is above the max
cardinality.

##### What is happening?

When cost-attribution active-series tracker observes more than the
maxCardinality label combinations, it enters the overflow and tracks the
time when it entered the overflow mode.

After a cooldown period (20 minutes by default) we delete the tracker,
which causes active-series tracker to reload (in order to correctly fill
the cost atttribution active series tracker again), and this process
starts again.

#### How do we fix this?

We will track up to `2*maxCardinality` label combinations internally,
this way we can have some insights about the customer behavior, even
though it's not absolutely precise.

After the cooldown period, we'll check whether the number of label
combinations is below the `maxCardinality`, and if it's not, we'll wait
another cooldown period, and so on.

This way we will only reload the trackers if customer has removed at
least `maxCardinality` combinations and didn't add new ones. It is still
possible that the customer is over the limit after this, because we
can't know how much cardinality is added from the series that we've
tracked in the `overflowCounter`: that information is lost forever, but
a situation where there's cardinality in those ones and not in the
observed ones is unlikely.

#### Which issue(s) this PR fixes or relates to

Fixes an internal issue.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>