lookup planning: adjust plan's intersection cost
We've seen that ingesters can end up doing more work when this optimization is enabled. This is due to fetching significantly more series and checking them for sharding

this PR adjusts the cost for retrieving a single series. The cost of doing these depends a lot on whether the block is an in-memory or an on-disk block. In-memory blocks have much more efficient sharding code.

Potential next steps:
 1. is to detect whether the block is in-memory or on-disk and adjust our plans.
 2. adjust the cost based on whether we have query sharding enabled or not; hashing series is expensive

1. will probably give higher returns

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>