Remote execution: don't override provided error type, if the error has one (#12809)
#### What this PR does

This PR changes the behaviour of `Dispatcher` to respect the error type
in errors returned by methods called during processing, rather than
always using a particular error type based on which method returned the
error.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12302 and
https://github.com/grafana/mimir/pull/12551

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.