usagestats: report "ingest-storage" or "classic" architecture mode (#12753)
#### What this PR does

The PR updates the anonymous usage statistics, so it reported the Mimir
architecture (mode), the installation runs with.

Note, I've picked the key `mode`, and not "architecture" because the
stats already report "arch" for the CPU architecture. I find it
confusing. We already used "ingest-storage mode" in some context in the
docs and in changelog, so I find it fitting.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>