Fix bug with pruning toggles not working in query planner (#12783)
#### What this PR does

Pruning query toggles was broken with
https://github.com/grafana/mimir/pull/12303 when it moved to AST
optimization passes, as the query planner preprocess step wraps some
parts of the expression in StepInvariantExpressions which we did not
account for in the code and unit tests working on it in isolation
instead of as part of the query plan.

This fixes that bug and adds preprocessing to the unit tests for some
AST optimization passes. Will do testing with more of the query plan in
another [PR](https://github.com/grafana/mimir/pull/12784) as that
requires bigger changes to the planning code.

#### Which issue(s) this PR fixes or relates to

Follow up to https://github.com/grafana/mimir/pull/12303

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>