jsonnet: Allow the max OTEL events per span to be configured (#12865)
#### What this PR does

Allow the max number of events in a span to be configured. We sometimes
hit the default limit of 128 events leading to earlier events being
dropped from the span.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>