Remote execution: don't panic if stream ends unexpectedly (#12870)
#### What this PR does

This PR fixes an issue where query-frontends can panic if the stream
from queriers ends unexpectedly.

This shouldn't be happening, but if it does happen for some reason, then
we don't want the entire process to crash.

#### Which issue(s) this PR fixes or relates to

#12551

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.