chore(deps): Update dskit support additional cluster validation labels (#12850)
#### What this PR does

Updates dskit to support additional cluster validation labels. The
additional labels are used on the server side and particular for the
query-frontend to accept requests from a remote cluster. This scenario
comes into play during a cluster migration where a target cluster rulers
needs to query the source query frontend for samples stored on the
source ingesters.

Commits in this branch:
- e57852904f chore(deps): Update dskit support additional cluster
validation labels

#### Which issue(s) this PR fixes or relates to


#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.