distributor: Report correct size in "message too big" error (#12799)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

In the Prometheus push endpoint, when the write request _after_
compression exceeds the limit, the HTTP request's content length (this
_before_ compression) would be reported, so you could see an error like
this:

```
the incoming push request has been rejected because its message size of 16245 bytes is larger than the allowed limit of 100000 bytes
```

With this change:

* Compressed and uncompressed sizes are now handled explicitly.
* The error message always has either `(compressed)` or `(uncompressed)`
to make it always clear which one it's talking about.
* The implementation of the several "message too large" error types have
been made uniform.

#### Which issue(s) this PR fixes or relates to

—

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>