[Docs] Update deployment modes topic (#12501)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

This PR updates the [deployment modes
documentation](https://grafana.com/docs/mimir/latest/references/architecture/deployment-modes/)
topic with regard to ingest storage architecture in Grafana Mimir.

It performs a general clean up of the file for style consistency and
clarity.

It also marks that the "read-write" deployment mode is not supported in
Mimir 3.0.

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir-squad/issues/3140

#### Checklist

- [ ] Tests updated.
- [X] Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.