Remote execution: fix two cancellation-related issues (#12856)
#### What this PR does

This PR fixes two issues in the query-frontend:

##### Unnecessary request cancellation messages are sometimes sent to
the query-scheduler

This PR fixes a race condition where queries evaluated with remote
execution could sometimes result in a cancellation message being sent to
the query-scheduler after the response had been completely read by
query-frontends.

This issue had no user-visible impact, but it causes confusing messages
in request traces, and in some cases can cause unnecessary shuffling of
queriers if shuffle sharding is enabled. (This happens because
cancelling the request causes the query-scheduler to close the
querier/query-scheduler stream, and if all streams for a single querier
are closed before any are re-established, then the querier is removed
from all shuffle shards and so reassignment of queriers to tenants
occurs.)

Relevant commits: all commits except 84b2bd4

##### `ProtobufResponseStream.Next()` sometimes returns a `nil` message
if the context passed to `DoProtobufRequest` is cancelled after at least
some of the response is received from the querier

This PR fixes an issue where `Next` might sometimes return a `nil`
message if the overall request context is cancelled after the
query-frontend has started to receive the response from the querier and
the caller has consumed all messages sent to the query-frontend. This
can result in errors like those
https://github.com/grafana/mimir/pull/12870 tries to mitigate.

Relevant commits: 84b2bd4

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12551

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.