Add support for query sharding inside MQE (#12835)
#### What this PR does

This PR implements support for query sharding inside MQE as an AST
optimisation pass.

There are some missing features that will be added in subsequent PRs:

* we need to buffer query responses for sharded legs in query-frontends
to avoid deadlocking if multiple query-frontends are evaluating
shardable queries at the same time and the total number of legs exceeds
the number of available querier workers:
https://github.com/grafana/mimir/pull/12885
* we need to limit the number of sharded legs for a single query request
evaluated at once, similar to the behaviour of the
`limitedParallelismRoundTripper`:
https://github.com/grafana/mimir/pull/12883
* we need to block user requests that use the `__sharded_concat__()`, as
this may otherwise lead to unpredictable behaviour:
https://github.com/grafana/mimir/pull/12886

:warning: The integration tests in this PR will only reliably pass with
the fix from https://github.com/grafana/mimir/pull/12856.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.