index planning: add configurable cost
Since our iteration speed is a bit limited, it would be helpful to tune the cost of the planner without having to recompile it. This PR extracts some threshold and costs as configration options.

```
-blocks-storage.tsdb.index-lookup-planning.min-series-per-block-for-query-planning
-blocks-storage.tsdb.index-lookup-planning.retrieved-posting-cost
-blocks-storage.tsdb.index-lookup-planning.retrieved-posting-list-cost
-blocks-storage.tsdb.index-lookup-planning.retrieved-series-cost
```

For consistency, I also renamed two existing configuration flags.

```
# PREVIOUSLY
-blocks-storage.tsdb.index-lookup-planning-comparison-portion
-blocks-storage.tsdb.index-lookup-planning-enabled

# NOW
-blocks-storage.tsdb.index-lookup-planning.comparison-portion
-blocks-storage.tsdb.index-lookup-planning.enabled
```

This PR is based on 12830 because of a config option added there

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

# Conflicts:
#	pkg/ingester/ingester.go