fix(parquet store gateway): map limit errors to UnprocessableEntity status (#12892)
Limit errors coming from parquet-common are currently mapped to Internal
errors in gRPC. Those will be retried by queriers:
https://github.com/grafana/mimir/blob/443ec0c7bb290e42e587a64697822e894cf96a28/pkg/querier/blocks_store_queryable.go#L1049-L1059


We should instead return a `422`, like the main path does: 

https://github.com/grafana/mimir/blob/443ec0c7bb290e42e587a64697822e894cf96a28/pkg/storegateway/limiter.go#L62-L73

A nicer way of fixing this would be introducing an interface in
parquet-common for the limiter, so we can return our own errors. The
simpler way I went with is by detecting the error type and wrapping it
with the correct status, at the top level to catch all paths.

See parquet common:
https://github.com/grafana/mimir/blob/443ec0c7bb290e42e587a64697822e894cf96a28/vendor/github.com/prometheus-community/parquet-common/search/limits.go#L34-L36