Fix issue where binary expressions with one side containing parentheses aren't sharded (#12900)
#### What this PR does

This PR fixes an issue where expressions like `sum(foo) or (sum(bar) *
2)` aren't sharded.

This is due to a bug in `canShardAllVectorSelectors` introduced in
https://github.com/grafana/mimir/pull/12677.

I haven't added a changelog entry given
https://github.com/grafana/mimir/pull/12677 isn't in the changelog and
that PR has not been included in a release yet.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12677

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.