Block-builder-scheduler: safe handling of multiple concurrent jobs in a partition (#12772)
#### What this PR does

Add a buffer of active completions where a job is only garbage collected
and used to advance the commit when both:
1. it is complete
2. all known jobs with a lower offset have both been completed and
garbage collected.

This is the remaining safety mechanism required for us to relax the
`max_jobs_per_partition` setting and consume jobs in parallel within a
partition.

There are two features of the scheduler that make use of this:

1. In normal operation, where the scheduler is creating new jobs,
assigning them, and workers are completing them.
1. At startup: in `finalizeObservations` we process observed job updates
and form an idea of planned and completed jobs. Any such observations
whose start and end offsets form a contiguous offset range starting from
the stored commit will be added (in ascending order) to the plannedJobs
structure. Further, any such jobs which are _completed_ in
`finalizeObservation` will have `partitionState.completeJob` called on
them, which contains the safe offset advancement/GCing mentioned
previously.

After we've amassed a certain level of confidence, the "pending" job
system and the job creation policy things can all be deleted, as they
only existed as a safety measure until we had this safe book-keeping
implemented.